Made the logic for evaluating the input expression:
In JavaScript, you can use either single quotes (') or double quotes (") to define a string. Both are valid ways to create strings, and they behave the same in most situations.

used try and catch if eval(input) throws an error:

For example: "2++3" will setInput as "Error"

If the input string represents a numerical expression, such as "2 + 3", the result will be a number.

const variable = 42;
console.log(typeof variable); // Output: "number"