State Variables:
todos: Manages the list of todo items.
inputValue: Keeps track of the value in the input field for adding new todos.
editMode: Tracks whether the application is in edit mode or not.
editId: Stores the ID of the todo item being edited.
editValue: Holds the updated value of the todo item being edited.