CSS Styling:
Tailwind CSS classes are used to style the components.
The main container (<main>) is styled to center its children vertically and horizontally.
The calculator is enclosed in a white card (<div>) with shadow and rounded corners.
BMI Categories:

A list of BMI categories is displayed along with their respective ranges.