Event Handlers:
handleClick: This function is called when a user clicks on the "Add to Cart" button in the Shop component. It checks if the item is already in the cart and displays a warning message if it is. Otherwise, it adds the item to the cart.
handleChange: This function is called when the user adjusts the quantity of an item in the Cart component. It finds the index of the item in the cart, updates its quantity, and then updates the cart state accordingly. If the quantity becomes zero, it removes the item from the cart.